Fix missing win/lose images bug
The delay at the end of the draw function after the win or lose
image was drawn was preventing the draw function from finishing,
and as a result preventing the next frame from being rendered.
Adding a restartGame boolean and moving the delay code to the start
of the draw function fixed it.